+++ src/App/AppDelegate.swift
@@ -1,0 +1,18 @@
+import AppKit
+
+final class AppDelegate: NSObject, NSApplicationDelegate {
+    private var mainWindowController: MainWindowController?
+
+    func applicationDidFinishLaunching(_ notification: Notification) {
+        setupMainMenu()
+
+        mainWindowController = MainWindowController()
+        mainWindowController?.showWindow(nil)
+
+        NSApp.activate(ignoringOtherApps: true)
+    }
+
+    func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
+        return true
+    }
+}